pyproject.toml
[project]
name = "product-ideation-system"
version = "1.0.0"
description = "AI-powered product concept generation and market validation with automated social media export"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "Product Ideation System"}
]
keywords = ["ai", "product-management", "market-research", "ideation", "pmf", "social-media"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "langchain-core>=0.3.79",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.10",
    "matplotlib>=3.10.7",
    "pillow>=11.3.0",
    "pydantic>=2.12.2",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "seaborn>=0.13.2",
    "typer>=0.19.2",
]

[project.scripts]
product-ideation = "src.main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
